[Core] Split filesystem class declarations into their own header
* Added filesystem TDL and generated header for File and StorageDevice definitions
* Moved file event and date definitions into shared core IDL inputs
* Updated module sources to include the filesystem header where file APIs are used
