+++ src/core/CMakeLists.txt
@@ -20,16 +20,13 @@ set (FUNCTIONS  # Full paths are required for add_custom_command() dependencies 
    "${CMAKE_CURRENT_SOURCE_DIR}/fs_volumes.cpp")
 
 set (CORE_CLASSES
-   "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_file.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_metaclass.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_module.cpp"
-   "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_storagedevice.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_task.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_thread.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_time.cpp")
 
 SET (CORE_DEFS
-   "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_file_def.c"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_task_def.c"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_thread_def.c")
 
@@ -38,10 +35,8 @@ if (KOTUKU_DEFS_AVAILABLE)
    set (CORE_OUTPUTS
       ${CORE_DEFS}
       "${DOCS}/modules/core.xml"
-      "${DOCS}/modules/classes/file.xml"
       "${DOCS}/modules/classes/metaclass.xml"
       "${DOCS}/modules/classes/module.xml"
-      "${DOCS}/modules/classes/storagedevice.xml"
       "${DOCS}/modules/classes/task.xml"
       "${DOCS}/modules/classes/thread.xml"
       "${DOCS}/modules/classes/time.xml"
@@ -117,6 +112,21 @@ idl_gen ("defs/compression.tdl"
       "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
       "${CMAKE_CURRENT_SOURCE_DIR}/compression/class_compression.cpp"
       "${CMAKE_CURRENT_SOURCE_DIR}/compression/class_compressed_stream.cpp")
+
+# File class
+
+idl_gen ("defs/filesystem.tdl"
+   NAME filesystem_defs
+   OUTPUT "${INCLUDE_OUTPUT}/modules/filesystem.h"
+   BYPRODUCTS
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_file_def.c"
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_storagedevice_def.c"
+      "${DOCS}/modules/classes/file.xml"
+      "${DOCS}/modules/classes/storagedevice.xml"
+   DEPENDS
+      "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_file.cpp"
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_storagedevice.cpp")
 
 if (NOT MSVC)
    # A bug in (MinGW?) GCC requires the following flag is applied to ensure that __packed__ works correctly.
