+++ src/image/image.cpp
@@ -39,6 +39,7 @@ rendered image size.
 #include <kotuku/main.h>
 #include <kotuku/modules/image.h>
 #include <kotuku/modules/display.h>
+#include <kotuku/modules/filesystem.h>
 #include <kotuku/strings.hpp>
 #include "../link/linear_rgb.h"
